Add validation deleteCommentHandler

+++ Forum-Project/src/components/SingleComment/SingleComment.jsx
@@ -2,22 +2,24 @@ import PropTypes from "prop-types";
 import { useNavigate } from "react-router-dom";
 import { deleteComment } from "../../services/comments.services";
 import { toast } from "react-toastify";
+import { useContext } from "react";
+import { AuthContext } from "../../context/authContext";
 
 const SingleComment = (props) => {
-
+  const { userData } = useContext(AuthContext);
   const comment = props.value
-  
+  const commentAuthor = comment.userName
   const navigate = useNavigate();
 
-  const deletePostHandler = () => {
-    //if (true) {
+  const deleteCommentHandler = () => {
+    if (commentAuthor === userData.username) {
       deleteComment(comment.id).then(() => {
         navigate("/home");
         toast("You comment deleted permanently!");
-        console.log("You comment deleted permanently!")
+
       }).catch((error) => console.error(error));
 
-    //} else { toast('Only author can delete the post!') }
+    } else { toast.error('Only author can delete the comment!') }
   };
 
   return (
@@ -30,7 +32,7 @@ const SingleComment = (props) => {
           liked {'98'}
         </button>
         <button className=" rounded-full bg-gray-50 px-3 py-1.5 font-medium text-gray-600 hover:bg-gray-100"
-        onClick={deletePostHandler}>
+          onClick={deleteCommentHandler}>
           Delete Post
         </button>
       </div>
